Added src parameter for pip in python::requirements
A requirements file can have --editable resources; in such a case the
pip default is to install them in ~/src/, which can be unsuitable; for
example, if run by root, /root/src/ is usually unreadable by other
users. pip provides an --src parameter to specify an alternative
directory. This is useful for installing requirements systemwide,
outside a virtualenv.