Update virtualenv.pp
This class should use a file type to create the venv dir, otherwise the directory is not created (and thus the hole thing is failing). Example: you let run the exec as a "normal" user and want the dir to reside under /var/lib. Most certainly the provided user may not write there. With the file definition, the directory gets created by puppet user (usually root) and chown'ed afterwards.

Maybe it would be a good idea to use "ensure_resource" from stdlib to have the provided user/group verified to exist.