Make the inclusion of the epel class optional.
Without this patch, people in secure environments who are not directly
routed to the internet or who mirror their own EPEL repositories cannot
use this module as it currently mandates the usage of the epel module.
This patch makes that optional and defaults to including it, such that
it will continue to work as expected for current users.