include python for access to exec_prefix
According to puppet scope documentation: "availability of out-of-scope variables is evaluation order dependent. You should only access out-of-scope variables if the class accessing them can guarantee that the other class is already declared, usually by explicitly declaring it with include before trying to read its variables."

This means if python::virtualenv is used without loading the class python prior and strict variables is activated, then an error is thrown that python::exec_prefix does not exist.