Convert root group identifier to numeric
Without this change, the module fails to function in a variety of ways
when attempted on a non-Linux system.  Most other *nixes use some other
name for the root group, like wheel.  To avoid casing all over the place
for this, and at the same time to allow this module to function on
non-Linux, using the numeric group identifier of zero captures the
intent and the outcome where the group 'root' was used.