Add basic support for FreeBSD
This work updates the code to function on FreeBSD in a reasonably sane
manner.  This include both python2 and python3, and accounts for some
oddities of how Python is managed on the FreeBSD OS.

Some things have not been tested.  What has been tested functional here
are the following:

  * Basic installation of python2 and python3
  * Pip installation on python2 and python3
  * Virtualenv installation on python2 and python3
  * Virtualenv creation
  * Installing a package with pip inside a virtualenv

This means that some of the extra features of this module, like
gunicorn, the requirements etc, have not been tested as part of this
work.