Fix pip package name on EL7 (#347)
Now uses the correct package name 'python2-pip'.

Due to a recent change in EPEL the python-pip package was renamed
to python2-pip. This currently breaks this module on CentOS 7. This
commit makes sure that it defaults to python2-pip if the system is
running Python 2 on CentOS 7 when using EPEL.

* Remove the unneeded dependency on EPEL

* Added pip package tests on RHEL 5 and 6