ignore old Gentoo factsets with legacy facter
facterdb contains a very old factset for Gentoo with Facter version 1.7.
This version is very old and won't work at all with the current version
of the module. Since we don't set a specific version for Gentoo in the
metadata.json (because it's a rolling release distribution),
rspec-puppet-facts will even pick those very old factsets. We are
ignoring them now. facterdb contains more modern factsets so Gentoo is
still tested.