Use Python 3 for virtualenv test
Python 2 has been deprecated, we should default to testing with Python 3

In commit 04ebe2d16d1d408f9c5d049aaa0426dce1c7bbe2, distribute was added
as an option to be set to false if setting up a python3 virtualenv.
Therefore, set distribute to false because we have defaulted to py3.