Allow to skip management of python dev package
Just like we can skip managing the pip and venv packages, allow to
opt-out of managing the dev package.

Enforcing a state of these packages sometimes cause trouble if the user
of the module does not care about them but they get installed as another
package dependency (e.g. syslog-ng-mod-python depends on python-venv, by
default python-venv is ensured absent, so on each run Puppet wants to
install syslog-ng-mod-python (which install python-venv as a dependency)
or remove python-venv (which remove syslog-ng-mod-python as a
dependency)).  This can be avoided using the corresponding
`manage_XXX_package` paramater, but `manage_dev_package` was missing.