Add fallback on the default system version of Python
When Python is not installed, the python_version fact is undef, but we
need to know which version of Python is going to be installed by the
module in order to use the proper command invocation for building
virtual envs.

This issue only occurs when the user use the system version of Python,
so add the default version of Python for supported operating systems
into Hiera and use that as a fallback when the version of Python is
unknown.