Do not pass `--log` to `pip install`
When passing a `--log` option to `pip install`, _verbose_ logging is
appended to the specified file, and the default logging captured by
puppet is _quitter_ than the default.

When an installation error occur, the quieter output does not contain
the reason for the failure, and inspection of the log file is necessary
to determine the reason of the error.  This is impractical in CI systems
where this file cannot be retrieved easily, and makes fixing regressions
harder.

On production systems, this log is also mostly useless since projects
dependencies versions are generally expected to be found in the
requirement file used for installation in order for deployments to be
reproducible.